# intrackperformanceAppraisal/intrackproject/openapi/openapi/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8081
paths:
  /task:
    post:
      operationId: createtaskcompltion
      parameters:
      - explode: true
        in: query
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompletion'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskCompletion'
          description: OK
      tags:
      - task-completion-controller
      x-content-type: application/json
      x-accepts:
      - '*/*'
  /interns:
    get:
      operationId: getAllInterns
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/Intern'
                type: array
          description: OK
      tags:
      - intern-controller
      x-accepts:
      - '*/*'
    post:
      operationId: createIntern
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Intern'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Intern'
          description: OK
      tags:
      - intern-controller
      x-content-type: application/json
      x-accepts:
      - '*/*'
  /feedback:
    post:
      operationId: createFeedbacks
      parameters:
      - explode: true
        in: query
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Feedback'
              type: array
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/Feedback'
                type: array
          description: OK
      tags:
      - feedback-controller
      x-content-type: application/json
      x-accepts:
      - '*/*'
  /attendance:
    post:
      operationId: createAttendance
      parameters:
      - explode: true
        in: query
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attendance'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Attendance'
          description: OK
      tags:
      - attendance-controller
      x-content-type: application/json
      x-accepts:
      - '*/*'
  /task/{internId}:
    get:
      operationId: getTaskCompletionByInternId
      parameters:
      - explode: false
        in: path
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskCompletionDto'
          description: OK
      tags:
      - task-completion-controller
      x-accepts:
      - '*/*'
  /interns/{internId}:
    delete:
      operationId: deleteIntern
      parameters:
      - explode: false
        in: path
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - intern-controller
      x-accepts:
      - application/json
    get:
      operationId: getInternById
      parameters:
      - explode: false
        in: path
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Intern'
          description: OK
      tags:
      - intern-controller
      x-accepts:
      - '*/*'
  /interns/{internId}/history:
    get:
      operationId: getInternHistory
      parameters:
      - explode: false
        in: path
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternHistoryDTO'
          description: OK
      tags:
      - intern-controller
      x-accepts:
      - '*/*'
  /feedback/{internId}:
    get:
      operationId: getFeedbackByInternId
      parameters:
      - explode: false
        in: path
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FeedbackDto'
          description: OK
      tags:
      - feedback-controller
      x-accepts:
      - '*/*'
  /attendance/{internId}:
    get:
      operationId: getAttendanceByInternId
      parameters:
      - explode: false
        in: path
        name: internId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AttendanceDto'
          description: OK
      tags:
      - attendance-controller
      x-accepts:
      - '*/*'
components:
  schemas:
    Intern:
      example:
        name: name
        id: 6
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
      type: object
    TaskCompletion:
      example:
        intern:
          name: name
          id: 6
        taskName: taskName
        completionDate: 2000-01-23
        id: 0
        completed: true
      properties:
        id:
          format: int64
          type: integer
        taskName:
          type: string
        completed:
          type: boolean
        completionDate:
          format: date
          type: string
        intern:
          $ref: '#/components/schemas/Intern'
      type: object
    Feedback:
      example:
        intern:
          name: name
          id: 6
        rating: 6
        comment: comment
        id: 0
      properties:
        id:
          format: int64
          type: integer
        comment:
          type: string
        rating:
          format: int32
          type: integer
        intern:
          $ref: '#/components/schemas/Intern'
      type: object
    Attendance:
      example:
        date: 2000-01-23
        intern:
          name: name
          id: 6
        id: 0
        present: true
      properties:
        id:
          format: int64
          type: integer
        date:
          format: date
          type: string
        present:
          type: boolean
        intern:
          $ref: '#/components/schemas/Intern'
      type: object
    TaskCompletionDto:
      example:
        taskCompletionList:
        - intern:
            name: name
            id: 6
          taskName: taskName
          completionDate: 2000-01-23
          id: 0
          completed: true
        - intern:
            name: name
            id: 6
          taskName: taskName
          completionDate: 2000-01-23
          id: 0
          completed: true
        completionPercentage: 0.8008281904610115
      properties:
        completionPercentage:
          format: double
          type: number
        taskCompletionList:
          items:
            $ref: '#/components/schemas/TaskCompletion'
          type: array
      type: object
    AttendanceDto:
      example:
        attendancePercentage: 0.8008281904610115
        attendanceList:
        - date: 2000-01-23
          intern:
            name: name
            id: 6
          id: 0
          present: true
        - date: 2000-01-23
          intern:
            name: name
            id: 6
          id: 0
          present: true
      properties:
        attendanceList:
          items:
            $ref: '#/components/schemas/Attendance'
          type: array
        attendancePercentage:
          format: double
          type: number
      type: object
    FeedbackDto:
      example:
        averageRating: 6.027456183070403
        feedbackList:
        - intern:
            name: name
            id: 6
          rating: 6
          comment: comment
          id: 0
        - intern:
            name: name
            id: 6
          rating: 6
          comment: comment
          id: 0
      properties:
        feedbackList:
          items:
            $ref: '#/components/schemas/Feedback'
          type: array
        averageRating:
          format: double
          type: number
      type: object
    InternHistoryDTO:
      example:
        intern:
          name: name
          id: 6
        feedbacks:
          averageRating: 6.027456183070403
          feedbackList:
          - intern:
              name: name
              id: 6
            rating: 6
            comment: comment
            id: 0
          - intern:
              name: name
              id: 6
            rating: 6
            comment: comment
            id: 0
        taskCompletions:
          taskCompletionList:
          - intern:
              name: name
              id: 6
            taskName: taskName
            completionDate: 2000-01-23
            id: 0
            completed: true
          - intern:
              name: name
              id: 6
            taskName: taskName
            completionDate: 2000-01-23
            id: 0
            completed: true
          completionPercentage: 0.8008281904610115
        attendanceRecords:
          attendancePercentage: 0.8008281904610115
          attendanceList:
          - date: 2000-01-23
            intern:
              name: name
              id: 6
            id: 0
            present: true
          - date: 2000-01-23
            intern:
              name: name
              id: 6
            id: 0
            present: true
      properties:
        intern:
          $ref: '#/components/schemas/Intern'
        attendanceRecords:
          $ref: '#/components/schemas/AttendanceDto'
        taskCompletions:
          $ref: '#/components/schemas/TaskCompletionDto'
        feedbacks:
          $ref: '#/components/schemas/FeedbackDto'
      type: object